requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
clarifai>=9.0.0
streamlit>=1.31.0
python-multipart>=0.0.5
google-search-results>=2.4.2  # For SerpAPI
newspaper3k>=0.2.8  # For article extraction
lxml[html_clean]>=4.9.0  # Required by newspaper3k
requests-cache>=1.2.0  # For API response caching

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.10.1
flake8>=6.0.0
